requirements.txt
torch
torchvision
diffusers
transformers
opencv-python
numpy
scikit-image
Pillow
